# functions/openapi.yaml
openapi: 3.1.0
info:
  title: LumiMD API
  version: 1.0.0
  description: |
    Backend API for LumiMD - a patient-facing medical navigation app.
    
    ## Authentication
    Most endpoints require a Firebase ID token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Error Format
    All errors follow this format:
    ```json
    {
      "code": "error_code",
      "message": "Human readable message",
      "details": { "field": "optional context" }
    }
    ```
servers:
  - url: https://us-central1-lumimd.cloudfunctions.net/api
    description: Production
  - url: https://us-central1-lumimd-dev.cloudfunctions.net/api
    description: Development
...
paths:
...
  /v1/visits:
    post:
      tags: [visits]
      summary: Start a new visit
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startedAt]
              properties:
                startedAt:
                  type: string
                  format: date-time
                  example: "2024-01-15T10:30:00Z"
      responses:
        '201':
          description: Visit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [visits]
      summary: List user visits
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [recording, completed]
      responses:
        '200':
          description: List of visits
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Visit'
  /v1/visits/{id}:
    get:
      tags: [visits]
      summary: Get visit detail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Visit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Actions ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [unauthorized, forbidden, not_found, validation_failed, rate_limited, conflict, server_error]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Visit:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [recording, completed]
        audioPath:
          type: string
        transcriptPath:
          type: string
        summary:
          $ref: '#/components/schemas/VisitSummary'
    VisitSummary:
      type: object
      properties:
        chiefConcern:
          type: string
        assessment:
          type: string
        plan:
          type: string
        redFlags:
          type: array
          items:
            type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Missing or invalid authorization header"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "not_found"
            message: "Resource not found"
...